refactor: 리팩토링 중 변경된 함수명 등 수정 및 불필요한 파일 삭제
